Always emit line break BreakPhrase when processing a <br/>
The following pattern would previously resolve the BR into a single
whitespace:

```
<p>
  <span>This is the first line.</span>
  <br/>
  <span>This is the second line.</span>
</p>
```

This is because, the "\n  " text element would be concatenated with
another "\n" from the `<br/>` element, then concatenated with the
following "\n  " text into a single text phrase. The default text
processing rules allow for TTML allow for mid-p repetitive whitespace
text entities to be interpreted as a single space.

Change
========
When the PhraseCollector gets a `<br/>` element, we emit the previous
phrase (if needed) and emit a BreakPhrase instance of type "line break".
This tells the paragraph collector emit a paragraph.