Fix upstream TTT issue #166, relative font sizes being re-computed if inherited
For instance, if a font size is specified as "75%", then it should be set to
the 75% of the font size of the parent element. But if that font size of 75%
is inherited by a child element, then it should not be 75% of 75%, it should
still just be 75%.

This commit adds a general-purpose system for marking style settings
(StyleSpecification class) as inherited. This way, settings can be treated
differently if inherited from a parent element vs. set directly on the current
element.

The `inherited` property is determined at initial TTML parse time. However,
during processing, as an intermediate step, the TTML document is converted to
the ISD XML representation. So code was required to encode and then retrieve
the `inherited` property for each style setting to/from the ISD
representation. This required adding a new namespace for vendor-specific
extensions to the ISD schema.