fix textEmphasis on a <span> that contains more nested <span>s
This fixes an unimplemented TODO comment. The code that handles <span>
elements with `tts:textEmphasis` applied (`AnnotatedPhraseCollector`) can now
recursively create a new `AnnotatedPhraseCollector` if it contains nested
<span> elements. Required changing where the collected elements are
accumulated-- because multiple recursive `AnnotatedPhraseCollector` objects
can now exist, the `AnnotatedPhraseCollector` itself is no longer a suitable
home for the accumulator.

The fix for TTT issue #166 had affected the ISD representation of one of the
unit tests, causing some of its elements to become more deeply nested, and
requiring that this unimplemented TODO be implemented as well.